requirements.txt
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
plotly==5.8.0
seaborn==0.11.2
streamlit==1.9.2
